Initial commit!
I recently became an engineering manager at PlanGrid (recently
acquired by Autodesk). As part of that project we had an informal
Management 101 session, which introduce the concept of a manager's
toolkit to me. A toolkit is a great way to codify your processes, is
something you can look back on, and is something you can take with you
when moving jobs.

I've since been playing around with generating my own toolkit. Also,
it was put to me that I should make a blog out of it, as a way to
share my experiences/learnings, but also to elicit feedback as another
way of rounding out my knowledge.

Anyway, here's my first commit. This toolkit is to be run as a GitHub
wiki, with commit messages to be presented as blog posts. That way
each post will tie back to changes in the wiki, listing what what's
changed and why.